Remove unnecessary resources from kubebuilder
This CRD is not meant to be deployed - the CRD yaml itself could be
useful just to see what is happening from kubebuilder comments, but it
should never need the webhook patch or anything like that.

Other details from kubebuilder that we don't plan on using have also
been removed to keep things simple.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>