Implement initial core spec and status
The fields added are present in all existing policy controllers.

The NamespaceSelector is implemented differently than previously, but
captures the previous behavior and extends it. The fuzz tests were
partly just for practice.

IsEnforce and IsInform are methods as opposed to (effectively) using an
enum, because that would basically require/enforce a consistent
capitalization, which could be a change in behavior. This has the
potential to be extended in the future.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>