Core types (#2)
* Add linting configuration

The linter configurations are largely based on what is in
github.com/open-cluster-management-io/governance-policy-addon-controller
but trimmed so that default behavior is generally not specifically
configured.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Kubebuilder create api PolicyCore

Command used:
```shell
kubebuilder create api --group policy --version v1beta1 \
  --kind PolicyCore --resource=true --controller=false
```

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Remove unnecessary resources from kubebuilder

This CRD is not meant to be deployed - the CRD yaml itself could be
useful just to see what is happening from kubebuilder comments, but it
should never need the webhook patch or anything like that.

Other details from kubebuilder that we don't plan on using have also
been removed to keep things simple.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Implement initial core spec and status

The fields added are present in all existing policy controllers.

The NamespaceSelector is implemented differently than previously, but
captures the previous behavior and extends it. The fuzz tests were
partly just for practice.

IsEnforce and IsInform are methods as opposed to (effectively) using an
enum, because that would basically require/enforce a consistent
capitalization, which could be a change in behavior. This has the
potential to be extended in the future.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>