Add a test for NamespaceSelector errors (#5)
* Fix typos and autoformatting

Sometimes the kubebuilder comment tags get re-formatted with an extra
space... this seems to depend on the whitespace around the comments, so
the new newlines here seem to fix it.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Update dependencies

Configuration matching stolostron/governance-policy-framework

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Add a fuzz-test target

The fuzz tests don't run during the normal `go test ./...` command, and
it can be slightly tricky to remember how to run the fuzz tests. This
target is not run in the CI.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Add test for NamespaceSelector errors

The fuzz tests actually cover this better, since they help verify that
the error always matches the expected kind, but these new tests can be
run in the CI and provide measurable code coverage.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

---------

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>