Update fakepolicy main to be testable
Now `Run` takes a context and a client config, and returns an error.
These can be provided and checked by the test suite. When the `cfg` is
not provided, it will get one from the controller-runtime. Some defaults
were updated to not listen on the network, so the tests don't ever need
those permissions.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>