Implement LabelSelector for NamespaceSelector
This takes a slightly different approach than what is in the config-
policy-controller, by embedding the LabelSelector type. Tests verify
that it was not necessary to put that type's fields explicitly in the
selector as pointers - Include and Exclude still operate independently
of the LabelSelector here.

A new Target type was added, which can be used for a generalization of
the NamespaceSelector idea.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>